fixed imports, moved more into validations, organized code into logical files
