--- 5
@@ -1,4 +1,0 @@
-[backend-db 011610a] user validations, changed user model email length -
- 3 files changed, 42 insertions(+), 31 deletions(-)
- create mode 100644 5
- rewrite backend/routes/api/users.js (64%)
