+++ backend/routes/api/users.js
@@ -3,22 +3,21 @@
 // Authentication and authorization are handled in the routes/api/auth.js file.
 const express = require('express');
 const router = express.Router();
+const {  restoreUser } = require('../utils/auth');
+const { signup } = require('../../controllers/api/userController');
+const { validateSignup } = require('../utils/validation');
 
-const { setTokenCookie, restoreUser } = require('../../utilities/auth');
-const { handleValidationErrors, } = require('../utils/validation');
-const { signup } = require('../../controllers/utils/userController');
-const { validateSignup } = require('../utils/userValidations');
 
 // Sign-up new User
 router.post('/', validateSignup, signup);
 
-// Get current user
-router.get('/', )
+// // Get current user
+// router.get('/', )
 
-// Edit current user
-router.put('/', )
+// // Edit current user
+// router.put('/', )
 
-// Delete current user
-router.delete('/', )
+// // Delete current user
+// router.delete('/', )
 
 module.exports = router;
