+++ backend/routes/index.js
@@ -4,6 +4,4 @@ const apiRouter = require('./api');
 
 router.use('/api', apiRouter)
 
-
-
 module.exports = router;
