--- value
@@ -1,3 +1,0 @@
-[backend-db 7bb7c23] sequelize notes added, and added user email validation and changed val -
- 4 files changed, 23 insertions(+), 4 deletions(-)
- create mode 100644 value
